extracted validatedoutlinedtextfieldparams
