+++ core/design/src/main/kotlin/edu/stanford/spezi/core/design/component/validated/outlinedtextfield/ValidatedOutlinedTextField.kt
@@ -1,19 +1,14 @@
-package edu.stanford.spezi.core.design.component
+package edu.stanford.spezi.core.design.component.validated.outlinedtextfield
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
@@ -21,34 +16,24 @@ import edu.stanford.spezi.core.design.theme.Colors
 import edu.stanford.spezi.core.design.theme.TextStyles.labelSmall
 
 @Composable
-fun ValidatedTextField(
+fun ValidatedOutlinedTextField(
     modifier: Modifier = Modifier,
     value: String = "",
     onValueChange: (String) -> Unit,
     labelText: String = "",
     errorText: String? = null,
-    singleLine: Boolean = true,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next),
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-    readOnly: Boolean = false,
-    trailingIcon: @Composable (() -> Unit)? = null,
 ) {
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
-        TextField(
+        OutlinedTextField(
             modifier = modifier.fillMaxWidth(),
             value = value,
             onValueChange = {
                 onValueChange(it)
             },
-            label = {
-                Text(labelText)
-            },
-            singleLine = singleLine,
-            keyboardOptions = keyboardOptions,
+            label = { Text(labelText) },
+            singleLine = true,
+            keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next),
             isError = errorText != null,
-            visualTransformation = visualTransformation,
-            trailingIcon = trailingIcon,
-            readOnly = readOnly
         )
         if (errorText != null) {
             Text(
@@ -60,39 +45,30 @@ fun ValidatedTextField(
     }
 }
 
+
 @Preview
 @Composable
-private fun ValidatedTextFieldPreview(
-    @PreviewParameter(ValidatedTextFieldProvider::class) params: ValidatedTextFieldParams
+private fun ValidatedOutlinedTextFieldPreview(
+    @PreviewParameter(ValidatedOutlinedTextFieldProvider::class) params: ValidatedOutlinedTextFieldParams
 ) {
-    ValidatedTextField(
-        value = params.value,
+    ValidatedOutlinedTextField(
         onValueChange = {},
+        value = params.value,
         labelText = params.labelText,
         errorText = params.errorText,
-        trailingIcon = {
-            IconButton(onClick = { }) {
-                Icon(Icons.Filled.Edit, contentDescription = "Select Date")
-            }
-        },
     )
 }
 
-private class ValidatedTextFieldProvider :
-    PreviewParameterProvider<ValidatedTextFieldParams> {
+private class ValidatedOutlinedTextFieldProvider :
+    PreviewParameterProvider<ValidatedOutlinedTextFieldParams> {
     override val values = sequenceOf(data, data.copy(errorText = null))
 }
 
 private val data =
-    ValidatedTextFieldParams(
+    ValidatedOutlinedTextFieldParams(
         value = "",
         labelText = "Label",
         errorText = "The input is invalid",
     )
 
 
-private data class ValidatedTextFieldParams(
-    val value: String,
-    val labelText: String,
-    val errorText: String?,
-)
