+++ core/design/src/main/kotlin/edu/stanford/spezi/core/design/component/validated/outlinedtextfield/ValidatedOutlinedTextFieldParams.kt
@@ -1,0 +1,7 @@
+package edu.stanford.spezi.core.design.component.validated.outlinedtextfield
+
+data class ValidatedOutlinedTextFieldParams(
+    val value: String,
+    val labelText: String,
+    val errorText: String?,
+)
